requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pinecone>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
requests>=2.31.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
